requirements.txt
faiss-cpu==1.7.4
ipykernel==6.23.1
ipython==8.13.2
jupyter_client==8.2.0
jupyter_core==5.3.0
numpy==1.24.3
openai==0.27.2
openapi-python-client==0.13.4
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
tqdm==4.65.0
typing_extensions==4.6.2